Simplify AddSupporterRatePlanItemToQueueLambda (don't worry about which items are batched together)
